+++ autoconfigure.sh
@@ -45,11 +45,14 @@ echo "[Successfully installed basic utilities for hyprland]"
 $AUR_MANAGER -S swww waybar waypaper aquamarine swaync nautilus btop htop hardinfo2 libnotify jq --noconfirm --needed
 echo "[Successfully installed file system management packages]"
 
-$AUR_MANAGER -S nwg-look nwg-dock-hyprland grim slurp wl-clipboard pacman-contrib --noconfirm --needed
+$AUR_MANAGER -S nwg-look nwg-dock-hyprland grim slurp wl-clipboard --noconfirm --needed
 echo "[Successfully installed dock for hyprland]"
 
 $AUR_MANAGER -S wl-clipboard qt5-wayland qt6-wayland qt6ct otf-font-awesome --noconfirm --needed
-echo "Finished installing hyprland configuration packages"
+echo "[Finished installing hyprland configuration packages]"
+
+$AUR_MANAGER -S gparted ncdu dysk pacman-contrib --noconfirm --needed
+echo "[Finished installing disk and cache management packages]"
 
 $AUR_MANAGER -S kitty zsh oh-my-posh-bin bash-completion \
    	zsh-completions fastfetch python-pywal postgresql --noconfirm --needed