+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 1. Added nvidia packages for egl
-2. Added paccache and pacman-contrib
+2. Added paccache(pacman-contrib) to handle cache bloat
+3. Added gparted, ncdu and dysk for disk management